+++ src_jni/ui_code/timelines.hh
@@ -74,8 +74,8 @@ private:
 	// END scale detector
 
 	struct CallbackContainer {
-		std::function<void(int, double, int, int)> cb;
-		CallbackContainer(std::function<void(int, double, int, int)>_cb) : cb(_cb) {}
+		std::function<void(double, double, int, int)> cb;
+		CallbackContainer(std::function<void(double, double, int, int)>_cb) : cb(_cb) {}
 	};
 
 	std::set<std::shared_ptr<CallbackContainer> > scroll_callbacks;
@@ -92,7 +92,7 @@ public:
 	TimeLines(KammoGUI::GnuVGCanvas* cnvs);
 	~TimeLines();
 
-	void add_scroll_callback(std::function<void(int, double, int, int)>);
+	void add_scroll_callback(std::function<void(double, double, int, int)>);
 
 	int get_sequence_minor_position_at(int horizontal_pixel_value);
 	int get_minor_spacing() {
